TYP: type-check against pandas-stubs; fix what it found
Add pandas-stubs to the lint group (and a docs group for the Sphinx
build). Fixes:
- zipline.testing.assert_timestamp_equal called the removed
  pandas.util.testing.assert_equal and always failed; it now compares
  directly (with tests).
- RollFinder stepped through sessions by `sessions.freq`, which is None
  for an index without a frequency; use the calendar's session offset.
- Portfolio.current_portfolio_weights is always float64 (an empty
  portfolio used to give an object-dtype Series).
- Use public pandas APIs instead of DatetimeIndex.asi8 and
  MultiIndex.set_levels round-trips; type HDF5 complib and the CLI
  timedelta unit.

+++ tests/test_algorithm.py
@@ -1013,7 +1013,7 @@ class TestPositions(zf.WithMakeAlgo, zf.ZiplineTestCase):
 
         expected_position_weights = [
             # No positions held on the first day.
-            pd.Series({}),
+            pd.Series({}, dtype="float64"),
             # Each equity's position value is its price times the number of
             # shares held. In this example, we hold a long position in 2 shares
             # of equity_1 so its weight is (95.0 * 2) = 190.0 divided by the
