+++ zipline/protocol.py
@@ -251,7 +251,8 @@ class Portfolio:
                     position.last_sale_price * position.amount * asset.price_multiplier
                 )
                 for asset, position in self.positions.items()
-            }
+            },
+            dtype="float64",
         )
         return position_values / self.portfolio_value
